Fix/main tag icon

+++ F1WM/ApiModel/News/NewsType.cs
@@ -2,7 +2,7 @@ namespace F1WM.ApiModel
 {
 	public class NewsType
 	{
-		public uint Id { get; set; }
+		public ushort Id { get; set; }
 		public string Title { get; set; }
 		public string AlternativeTitle { get; set; }
 	}
